requirements.txt
Pillow==10.4.0
numpy==1.26.4
PyQt5==5.15.10